// sql-movies/src/types.ts
export interface GenreRow {
  id: number;
  genre: string;
}

export interface ActorRow {
  id: number;
  full_name: string;
}

export interface DirectorRow {
  id: number;
  full_name: string;
}

export interface KeywordRow {
  id: number;
  keyword: string;
}

export interface ProductionCompanyRow {
  id: number;
  company_name: string;
}